+++ quilt-rs/src/fixtures.rs
@@ -18,15 +18,10 @@ pub mod manifest {
 
     use crate::Res;
 
-    const TEST_LOCAL_PARQUET: &str = "fixtures/manifest.parquet";
     const TEST_LOCAL_CHECKSUMMED: &str = "fixtures/checksummed.jsonl";
 
     pub const CHECKSUMMED_HASH: &str =
-        "9c4db11437f11c3bbe25b39601069b8ed09b39f5f18ac29a13df4361240859d9";
-
-    pub fn parquet() -> Res<PathBuf> {
-        local_uri(TEST_LOCAL_PARQUET)
-    }
+        "87c4c6caa2eaeb346ea8c3a5c3650a54b9ae93ed5229db67eac7449fc198f0c5";
 
     pub fn checksummed() -> Res<PathBuf> {
         local_uri(TEST_LOCAL_CHECKSUMMED)