+++ quilt-rs/src/manifest.rs
@@ -3,14 +3,9 @@
 
 #[allow(clippy::module_inception)]
 mod manifest;
-mod row;
-mod table;
+mod top_hasher;
 
-pub use row::Header;
-pub use row::Row;
-pub use row::RowDisplay;
-pub use table::Table;
-pub use table::TopHasher;
+pub use top_hasher::TopHasher;
 
 pub use manifest::Manifest;
 pub use manifest::ManifestHeader;